Restart service on persistent failure
This is a workaround for scapy's problem where it cannot always
successfully find the route for the packet.

Make the script exit with error if the route cannot be found and set the
service file to restart it when it fails. Upon restart, the error should
be resolved. This change should not be required if scapy found the route
as expected.